requirements.txt
cryptography==36.0.0
fastapi==0.70.0
PyJWT==2.3.0
PyMySQL==1.0.2
python-dotenv==0.19.2
SQLAlchemy==1.4.28
uvicorn==0.16.0
Werkzeug==2.0.2
